Clean up iOS code
