--- iosApp/Source/App/Sidebar.swift
@@ -1,9 +1,0 @@
-//
-//  Sidebar.swift
-//  FeedFlow
-//
-//  Created by Marco Gomiero on 19/10/23.
-//  Copyright © 2023 orgName. All rights reserved.
-//
-
-import Foundation
