--- iosApp/Source/HomeViewModel.swift
@@ -1,14 +1,0 @@
-//
-//  HomeViewModel.swift
-//  FeedFlow
-//
-//  Created by Marco Gomiero on 27/03/23.
-//  Copyright © 2023 orgName. All rights reserved.
-//
-
-import Foundation
-import shared
-import KMPNativeCoroutinesAsync
-
-// @MainActor
-// class HomeView
